feat: added countdown timer
